OGM-1002 Move MongoDB tests
