+++ integrationtest/src/test/java/org/hibernate/ogm/test/integration/mongodb/MongoDBModuleMemberRegistrationIT.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.test.integration.jboss;
+package org.hibernate.ogm.test.integration.mongodb;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -23,13 +23,14 @@ import org.hibernate.ogm.compensation.operation.CreateTuple;
 import org.hibernate.ogm.compensation.operation.GridDialectOperation;
 import org.hibernate.ogm.compensation.operation.OperationType;
 import org.hibernate.ogm.datastore.mongodb.MongoDB;
-import org.hibernate.ogm.test.integration.jboss.errorhandler.TestErrorHandler;
-import org.hibernate.ogm.test.integration.jboss.model.EmailAddress;
-import org.hibernate.ogm.test.integration.jboss.model.Member;
-import org.hibernate.ogm.test.integration.jboss.model.PhoneNumber;
-import org.hibernate.ogm.test.integration.jboss.service.ContactManagementService;
-import org.hibernate.ogm.test.integration.jboss.service.PhoneNumberService;
-import org.hibernate.ogm.test.integration.jboss.util.ModuleMemberRegistrationDeployment;
+import org.hibernate.ogm.test.integration.mongodb.errorhandler.TestErrorHandler;
+import org.hibernate.ogm.test.integration.mongodb.model.EmailAddress;
+import org.hibernate.ogm.test.integration.mongodb.model.PhoneNumber;
+import org.hibernate.ogm.test.integration.mongodb.service.ContactManagementService;
+import org.hibernate.ogm.test.integration.mongodb.service.PhoneNumberService;
+import org.hibernate.ogm.test.integration.testcase.ModuleMemberRegistrationScenario;
+import org.hibernate.ogm.test.integration.testcase.model.Member;
+import org.hibernate.ogm.test.integration.testcase.util.ModuleMemberRegistrationDeployment;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
