+++ integrationtest/src/test/java/org/hibernate/ogm/test/integration/mongodb/errorhandler/TestErrorHandler.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.test.integration.jboss.errorhandler;
+package org.hibernate.ogm.test.integration.mongodb.errorhandler;
 
 import java.util.ArrayList;
 import java.util.Collections;
