+++ integrationtest/src/test/java/org/hibernate/ogm/test/integration/mongodb/service/ContactManagementService.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.test.integration.jboss.service;
+package org.hibernate.ogm.test.integration.mongodb.service;
 
 import java.util.List;
 
@@ -12,8 +12,8 @@ import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 
-import org.hibernate.ogm.test.integration.jboss.model.EmailAddress;
-import org.hibernate.ogm.test.integration.jboss.model.PhoneNumber;
+import org.hibernate.ogm.test.integration.mongodb.model.EmailAddress;
+import org.hibernate.ogm.test.integration.mongodb.model.PhoneNumber;
 
 /**
  * @author Gunnar Morling
